Patch for IRLP SPEAKFREE_CNAME change
This changes bFoundCallsign in conference.c to initialize to TRUE instead of FALSE to work around a change IRLP made.  IRLP and EchoIRLP had been setting environment variable SPEAKFREE_CNAME = "CALLSIGN" when connecting to reflectors or experimental reflectors. Now IRLP sets it to the node call sign variable "${CALLSIGN}". theLinkbox and theBridge both look for the string "CALLSIGN" and if NOT found, clears the Callsign and Password variables. This resulted in IP addresses instead of node numbers being displayed from connected IRLP nodes on EchoLink nodes and tlb .users command output . Hoping Skip can provide some history on this behavior in tbd/tlb.  For now, I changed the boolean to avoid clearing the variables in tlb.